+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## 0.0.3
+
 ## 0.0.2
 
 * Plane detection enum implemented